requirements.txt
Flask==3.0.0
pymongo==4.6.0
PyJWT==2.8.0
bcrypt==4.1.1
flask-swagger-ui==4.11.1
requests==2.31.0
paramiko==3.3.1
